refactor(runtime): move observers together
