+++ packages/runtime/src/binding/let-binding.ts
@@ -1,10 +1,10 @@
 import { IIndexable, IServiceLocator, Reporter, Tracer } from '@aurelia/kernel';
 import { IBindScope, ILifecycle, State } from '../lifecycle';
 import { IScope, LifecycleFlags } from '../observation';
+import { IObserverLocator } from '../observation//observer-locator';
 import { IExpression } from './ast';
 import { IBindingTarget } from './binding';
 import { connectable, IConnectableBinding, IPartialConnectableBinding } from './connectable';
-import { IObserverLocator } from './observer-locator';
 
 const slice = Array.prototype.slice;
