+++ packages/runtime/src/index.ts
@@ -1,32 +1,32 @@
 
 
 
-
-export { ArrayObserver, enableArrayObservation, disableArrayObservation, nativePush, nativePop, nativeShift, nativeUnshift, nativeSplice, nativeReverse, nativeSort } from './binding/array-observer'; // TODO: do this differently, not let this be ugly, etc, etc
-export { MapObserver, enableMapObservation, disableMapObservation, nativeSet, nativeDelete as nativeMapDelete, nativeClear as nativeMapClear } from './binding/map-observer'; // TODO: do this differently, not let this be ugly, etc, etc
-export { SetObserver, enableSetObservation, disableSetObservation, nativeAdd, nativeDelete as nativeSetDelete, nativeClear as nativeSetClear } from './binding/set-observer'; // TODO: do this differently, not let this be ugly, etc, etc
 export * from './binding/ast';
 export * from './binding/binding-context';
 export * from './binding/binding-mode';
 export * from './binding/binding';
 export * from './binding/call';
-export * from './binding/collection-observer';
-export * from './binding/computed-observer';
-export * from './binding/dirty-checker';
-export * from './binding/element-observation';
-export * from './binding/event-manager';
 export * from './binding/expression-parser';
 export * from './binding/interpolation-binding';
 export * from './binding/let-binding';
 export * from './binding/listener';
-export * from './binding/observer-locator';
-export * from './binding/property-observation';
 export * from './binding/ref';
 export * from './binding/signaler';
-export * from './binding/subscriber-collection';
-export * from './binding/svg-analyzer';
-export * from './binding/target-accessors';
-export * from './binding/target-observer';
+
+export { ArrayObserver, enableArrayObservation, disableArrayObservation, nativePush, nativePop, nativeShift, nativeUnshift, nativeSplice, nativeReverse, nativeSort } from './observation/array-observer'; // TODO: do this differently, not let this be ugly, etc, etc
+export { MapObserver, enableMapObservation, disableMapObservation, nativeSet, nativeDelete as nativeMapDelete, nativeClear as nativeMapClear } from './observation/map-observer'; // TODO: do this differently, not let this be ugly, etc, etc
+export { SetObserver, enableSetObservation, disableSetObservation, nativeAdd, nativeDelete as nativeSetDelete, nativeClear as nativeSetClear } from './observation/set-observer'; // TODO: do this differently, not let this be ugly, etc, etc
+export * from './observation/collection-observer';
+export * from './observation/computed-observer';
+export * from './observation/dirty-checker';
+export * from './observation/element-observation';
+export * from './observation/event-manager';
+export * from './observation/observer-locator';
+export * from './observation/property-observation';
+export * from './observation/subscriber-collection';
+export * from './observation/svg-analyzer';
+export * from './observation/target-accessors';
+export * from './observation/target-observer';
 
 export * from './resources/binding-behavior';
 export * from './resources/binding-behaviors/attr';
