+++ packages/runtime/src/observation/svg-analyzer.ts
