+++ packages/runtime/src/resources/binding-behaviors/attr.ts
@@ -1,9 +1,9 @@
 import { IRegistry } from '@aurelia/kernel';
 import { Binding } from '../../binding/binding';
-import { DataAttributeAccessor } from '../../binding/target-accessors';
 import { IElement } from '../../dom.interfaces';
 import { ILifecycle } from '../../lifecycle';
 import { IScope, LifecycleFlags } from '../../observation';
+import { DataAttributeAccessor } from '../../observation/target-accessors';
 import { bindingBehavior } from '../binding-behavior';
 
 @bindingBehavior('attr')
