+++ packages/runtime/src/resources/custom-attributes/repeat.ts
@@ -2,11 +2,11 @@ import { IIndexable, inject, IRegistry } from '@aurelia/kernel';
 import { ForOfStatement } from '../../binding/ast';
 import { Binding } from '../../binding/binding';
 import { BindingContext, Scope } from '../../binding/binding-context';
-import { getCollectionObserver } from '../../binding/observer-locator';
-import { SetterObserver } from '../../binding/property-observation';
 import { IRenderLocation } from '../../dom.interfaces';
 import { IBindScope, IRenderable, IView, IViewFactory, State } from '../../lifecycle';
 import { CollectionObserver, IBatchedCollectionSubscriber, IObservedArray, IScope, LifecycleFlags, ObservedCollection } from '../../observation';
+import { getCollectionObserver } from '../../observation/observer-locator';
+import { SetterObserver } from '../../observation/property-observation';
 import { bindable } from '../../templating/bindable';
 import { ICustomAttribute, templateController } from '../custom-attribute';
