+++ packages/runtime/src/templating/lifecycle-render.ts
@@ -1,12 +1,12 @@
 import { all, Class, DI, IContainer, IDisposable, IIndexable, Immutable, ImmutableArray, inject, IRegistry, IResolver, IResourceDescriptions, PLATFORM, Registration, Reporter, RuntimeCompilationResources, Tracer, Writable } from '@aurelia/kernel';
 import { Scope } from '../binding/binding-context';
-import { Observer } from '../binding/property-observation';
-import { subscriberCollection } from '../binding/subscriber-collection';
 import { BindableDefinitions, buildTemplateDefinition, customElementBehavior, IHydrateElementInstruction, ITargetedInstruction, ITemplateDefinition, TemplateDefinition, TemplatePartDefinitions } from '../definitions';
 import { DOM, INodeSequenceFactory, NodeSequence, NodeSequenceFactory } from '../dom';
 import { IElement, INode, INodeSequence, IRenderLocation } from '../dom.interfaces';
 import { Hooks, ILifecycle, IRenderable, IRenderContext, IViewFactory } from '../lifecycle';
 import { IAccessor, IPropertySubscriber, ISubscribable, ISubscriberCollection, LifecycleFlags, MutationKind } from '../observation';
+import { Observer } from '../observation/property-observation';
+import { subscriberCollection } from '../observation/subscriber-collection';
 import { ICustomAttribute, ICustomAttributeType } from '../resources/custom-attribute';
 import { ICustomElement, ICustomElementType } from '../resources/custom-element';
 import { ViewFactory } from './view';
